migrada la tabla de activos con tests
