+++ pubspec.yaml
@@ -44,6 +44,7 @@ dependencies:
   textfield_tags:
     git:
       url: https://github.com/victorcatn/textfield_tags.git
+  uuid: ^3.0.5
 
 dev_dependencies:
   build_runner: ^2.1.4
